requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.2

# Code analysis
pylint>=3.0.0
flake8>=6.0.0
black>=23.0.0
mypy>=1.0.0
radon>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tenacity>=8.2.0
mcp>=1.0.0
google-adk>=0.1.0
# Web framework (for deployment)
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Jupyter for submission
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0


